+++ e2e/auth.e2e.ts
@@ -1,0 +1,5 @@
+describe("Authentication tests", () => {
+    test.todo("Identity token should alow to login with LoginStrategy");
+
+    test.todo("Delegate token should allow to login");
+});