+++ src/modules/assets/assets.service.ts
@@ -1,7 +1,6 @@
 import { utils, providers } from "ethers";
 import { Methods } from "@ew-did-registry/did";
 import { IdentityManager__factory } from "../../../ethers/factories/IdentityManager__factory";
-import { IdentityManager } from "../../../ethers/IdentityManager";
 import { OfferableIdentity__factory } from "../../../ethers/factories/OfferableIdentity__factory";
 import { ChainConfig, chainConfigs } from "../../config/chain.config";
 import { CacheClient } from "../cacheClient/cacheClient.service";
@@ -12,7 +11,9 @@ import { AssetHistoryEventType } from "./assets.types";
 export class AssetsService {
     private _owner: string;
     private _did: string;
-    private _assetManager: IdentityManager;
+    private _assetManager: string;
+    private _assetInterface = OfferableIdentity__factory.createInterface();
+    private _assetManagerInterface = IdentityManager__factory.createInterface();
 
     constructor(private _signerService: SignerService, private _cacheClient: CacheClient) {
         this._signerService.onInit(async () => await this.init());
@@ -30,18 +31,17 @@ export class AssetsService {
         this._did = `did:${Methods.Erc1056}:${this._owner}`;
         const chainId = (await signer.provider.getNetwork()).chainId;
         const chainConfig = chainConfigs()[chainId] as ChainConfig;
-        const { assetManagerAddress } = chainConfig;
-        this._assetManager = new IdentityManager__factory().attach(assetManagerAddress);
+        this._assetManager = chainConfig.assetManagerAddress;
     }
 
     async registerAsset(): Promise<string> {
-        const data = this._assetManager.interface.encodeFunctionData("createIdentity", [this._owner]);
-        const receipt = await this._signerService.send({ to: this._assetManager.address, data });
+        const data = this._assetManagerInterface.encodeFunctionData("createIdentity", [this._owner]);
+        const receipt = await this._signerService.send({ to: this._assetManager, data });
         const event = receipt.logs
-            .map((l) => this._assetManager.interface.parseLog(l))
+            .map((l) => this._assetManagerInterface.parseLog(l))
             .find(
                 (log) =>
-                    log.name === this._assetManager.interface.events["IdentityCreated(address,address,uint256)"].name,
+                    log.name === this._assetManagerInterface.events["IdentityCreated(address,address,uint256)"].name,
             ) as utils.LogDescription;
         const identity = event.args[0] as string;
         let asset = await this.getAssetById({ id: `did:ethr:${identity}` });
@@ -86,19 +86,19 @@ export class AssetsService {
     }
 
     async getOwnedAssets({ did = this._did }: { did?: string } = {}) {
-        return this._cacheClient.getOwnedAssets({ did });
+        return this._cacheClient.getOwnedAssets(did);
     }
 
     async getOfferedAssets({ did = this._did }: { did?: string } = {}) {
-        return this._cacheClient.getOfferedAssets({ did });
+        return this._cacheClient.getOfferedAssets(did);
     }
 
     async getAssetById({ id }: { id: string }) {
-        return this._cacheClient.getAssetById({ id });
+        return this._cacheClient.getAssetById(id);
     }
 
     async getPreviouslyOwnedAssets({ owner }: { owner: string }) {
-        return this._cacheClient.getPreviouslyOwnedAssets({ owner });
+        return this._cacheClient.getPreviouslyOwnedAssets(owner);
     }
 
     async getAssetHistory({
@@ -111,7 +111,7 @@ export class AssetsService {
         skip?: number;
         type?: AssetHistoryEventType;
     }) {
-        return this._cacheClient.getAssetHistory({ id, ...query });
+        return this._cacheClient.getAssetHistory(id, { ...query });
     }
 
     private offerAssetTx({
@@ -121,33 +121,29 @@ export class AssetsService {
         offerTo: string;
         assetContractAddress: string;
     }): providers.TransactionRequest {
-        const asset = new OfferableIdentity__factory().attach(assetContractAddress);
         return {
-            data: asset.interface.encodeFunctionData("offer", [offerTo]),
+            data: this._assetInterface.encodeFunctionData("offer", [offerTo]),
             to: assetContractAddress,
         };
     }
 
     private acceptOfferTx({ assetContractAddress }: { assetContractAddress: string }): providers.TransactionRequest {
-        const asset = new OfferableIdentity__factory().attach(assetContractAddress);
         return {
-            data: asset.interface.encodeFunctionData("acceptOffer"),
+            data: this._assetInterface.encodeFunctionData("acceptOffer"),
             to: assetContractAddress,
         };
     }
 
     private rejectOfferTx({ assetContractAddress }: { assetContractAddress: string }): providers.TransactionRequest {
-        const asset = new OfferableIdentity__factory().attach(assetContractAddress);
         return {
-            data: asset.interface.encodeFunctionData("rejectOffer"),
+            data: this._assetInterface.encodeFunctionData("rejectOffer"),
             to: assetContractAddress,
         };
     }
 
     private cancelOfferTx({ assetContractAddress }: { assetContractAddress: string }): providers.TransactionRequest {
-        const asset = new OfferableIdentity__factory().attach(assetContractAddress);
         return {
-            data: asset.interface.encodeFunctionData("cancelOffer"),
+            data: this._assetInterface.encodeFunctionData("cancelOffer"),
             to: assetContractAddress,
         };
     }