+++ src/utils/detectEnvironment.ts
@@ -1,27 +1,11 @@
-import detectEthereumProvider from "@metamask/detect-provider";
-
 export enum ExecutionEnvironment {
     NODE = "ExecutionEnvironment:Node",
     BROWSER = "ExecutionEnvironment:Browser",
 }
 
 export const executionEnvironment = () =>
-    (isNode() && ExecutionEnvironment.NODE) || (isBrowser() && ExecutionEnvironment.BROWSER);
+    isNode() && !isBrowser() ? ExecutionEnvironment.NODE : ExecutionEnvironment.BROWSER;
 
-const isNode = new Function("try { return this === global } catch(e) { return false; }");
+const isNode = () => typeof process !== "undefined" && process.versions != null && process.versions.node != null;
 
-const isBrowser = new Function("try { return this === window } catch(e) { return false; }");
-
-export const isMetamaskExtensionPresent = async () => {
-    const provider = (await detectEthereumProvider({ mustBeMetaMask: true })) as
-        | {
-              request: any;
-          }
-        | undefined;
-
-    const chainId = (await provider?.request({
-        method: "eth_chainId",
-    })) as number | undefined;
-
-    return { isMetamaskPresent: !!provider, chainId };
-};
+const isBrowser = () => typeof window !== "undefined" && typeof window.document !== "undefined";